rds: implement rds flavor
This is the first cut. RDS does not support GTIDs. However, we
can expect the master to not change over long periods of time.
This means that we can rely on file:position for tracking filtered
replication.

Since RDS runs mostly stock mysql. There's nothing in the handshake
that can be used to identify. Due to this, we rely on MYSQL_FLAVOR
being explicitly set to rds.

Other implemenation notes:
GTIDs are explicit events in the binlog stream, whereas file:pos
is implicit with every event. For now, the rds flavor creates pseudo
GTID events to cause the binlog streamer to update its position.

Also, 'SHOW SLAVE STATUS' could have a file:pos that is not at a
transaction boundary. This requires us to transmit non-transactional
events also to the targets. This is detected by checking the
autocommit flag. If it's on, then it's a standalone event and
the position should therefore be transmitted.

Signed-off-by: Sugu Sougoumarane <ssougou@gmail.com>